+++ lib/api/types/index.js
@@ -7,13 +7,9 @@ const {
 
 	SymbolSubtypeCollection,
 	SymbolConstructorName,
-	SymbolGaia,
 
 	TYPE_TITLE_PREFIX,
 	MNEMOSYNE,
-	MNEMONICA,
-	GAIA,
-	URANUS
 
 } = require('../../constants');
 
@@ -21,594 +17,22 @@ const {
 	ALREADY_DECLARED,
 	WRONG_TYPE_DEFINITION,
 	TYPENAME_MUST_BE_A_STRING,
-	WRONG_MODIFICATION_PATTERN,
 	HANDLER_MUST_BE_A_FUNCTION,
-	EXISTENT_PROPERTY_REDEFINITION,
 } = require('../../errors');
 
 const hooksApi = require('../hooks');
-const extract = require('../../utils/extract');
-const parent = require('../../utils/parent');
-const pick = require('../../utils/pick');
+const TypeProxy = require('./TypeProxy');
 
-const {
-	collectConstructors
-} = require('../../utils');
 
 const compileNewModificatorFunctionBody
 	= require('./compileNewModificatorFunctionBody');
 
-const
-	oldMC = require('./createInstanceModificator200XthWay'),
-	newMC = require('./createInstanceModificator');
-
-const getModificationConstructor = (useOldStyle) => {
-	return (useOldStyle ? oldMC : newMC)();
-};
-
-const getTypeChecker = (TypeName) => {
-	return (instance) => {
-		if (!instance) {
-			return false;
-		}
-		if (!instance.constructor) {
-			return false;
-		}
-		if (instance instanceof Promise) {
-			return instance[SymbolConstructorName] === TypeName;
-		}
-		const constructors = collectConstructors(instance);
-		return constructors[TypeName] || false;
-	};
-};
-
-const checkProto = (proto) => {
-	if (!(proto instanceof Object)) {
-		throw new WRONG_TYPE_DEFINITION('expect prototype to be an object');
-	}
-};
-
-const CreationHandler = function (constructionAnswer) {
-	// standard says : 
-	// if constructor returns something
-	// then this is a toy
-	// we have to play with
-	// respectively
-	// so we will not follow the rule
-	// if (constructionAnswer instanceof types[TypeName]) {
-	// and instead follow the line below
-	if (constructionAnswer instanceof Object) {
-		return constructionAnswer;
-	}
-	return this;
-};
-
-
-const addProps = function (proto) {
-
-	const {
-		type,
-		existentInstance,
-		args,
-		// } = opts;
-	} = this;
-
-	const {
-		namespace,
-		collection,
-		subtypes,
-	} = type;
-
-
-	odp(proto, '__proto_proto__', {
-		get() {
-			return proto;
-		}
-	});
-
-	odp(proto, '__args__', {
-		get() {
-			return args;
-		}
-	});
-
-	odp(proto, '__collection__', {
-		get() {
-			return collection;
-		}
-	});
-
-	odp(proto, '__namespace__', {
-		get() {
-			return namespace;
-		}
-	});
-
-	odp(proto, '__subtypes__', {
-		get() {
-			return subtypes;
-		}
-	});
-
-	odp(proto, '__type__', {
-		get() {
-			return type;
-		}
-	});
-
-	odp(proto, '__parent__', {
-		get() {
-			return existentInstance;
-		}
-	});
-
-	const timestamp = Date.now();
-	Object.defineProperty(proto, '__timestamp__', {
-		get() {
-			return timestamp;
-		}
-	});
-
-};
-
-
-const invokePreHooks = function (opts) {
-
-	const {
-		type,
-		existentInstance,
-		args,
-	} = opts;
-
-	const {
-		namespace,
-		collection,
-	} = type;
-
-	namespace.invokeHook('preCreation', {
-		type,
-		existentInstance,
-		args
-	});
-
-	collection.invokeHook('preCreation', {
-		type,
-		existentInstance,
-		args
-	});
-
-	type.invokeHook('preCreation', {
-		type,
-		existentInstance,
-		args
-	});
-
-};
-
-const invokePostHooks = function (inheritedInstance) {
-
-	const {
-		__type__: type,
-		__parent__: existentInstance,
-		__args__: args,
-	} = inheritedInstance;
-
-	const {
-		namespace,
-		collection,
-	} = type;
-
-	type.invokeHook('postCreation', {
-		type,
-		existentInstance,
-		inheritedInstance,
-		args
-	});
-
-	collection.invokeHook('postCreation', {
-		type,
-		existentInstance,
-		inheritedInstance,
-		args
-	});
-
-	namespace.invokeHook('postCreation', {
-		type,
-		existentInstance,
-		inheritedInstance,
-		args
-	});
-
-};
-
-
-const proceedInheritance = function (opts) {
-
-	const {
-		InstanceModificator,
-		args,
-	} = opts;
-
-	const inheritedInstance = new InstanceModificator(...args);
-	return inheritedInstance;
-
-};
-
-const addSubTypes = (inheritedInstance, subtypes) => {
-
-	const {
-		__proto_proto__: proto
-	} = inheritedInstance;
-
-	Object.entries(subtypes).forEach((subtype) => {
-
-		const [name, subType] = subtype;
-
-		if (inheritedInstance.hasOwnProperty(name)) {
-			throw new EXISTENT_PROPERTY_REDEFINITION(name);
-		}
-
-		if (!inheritedInstance.constructor) {
-			throw new WRONG_MODIFICATION_PATTERN('should inherit from mnemonica instance');
-		}
-
-		if (!inheritedInstance.constructor[SymbolConstructorName]) {
-			throw new WRONG_MODIFICATION_PATTERN('should inherit from mnemonica instance');
-		}
-
-		odp(proto, name, {
-
-			get() {
-
-				const SubTypeConstructor = function (...args) {
-
-					const inheritedSubInstance =
-						createInstance
-							.call(inheritedInstance, subType, args);
-
-					return inheritedSubInstance;
-
-				};
-
-				odp(SubTypeConstructor, Symbol.hasInstance, {
-					get() {
-						return getTypeChecker(subType.TypeName);
-					}
-				});
-
-				return SubTypeConstructor;
-
-			}
-
-		});
-
-	});
-};
-
-
-const Mnemosyne = function (namespace, gaia) {
-
-	const Mnemonica = function () {
-		odp(this, SymbolConstructorName, {
-			get() {
-				return namespace.name;
-			}
-		});
-	};
-
-	const mnemonica = Object.getPrototypeOf(gaia);
-
-	Object.setPrototypeOf(Mnemonica.prototype, mnemonica);
-	Mnemonica.prototype.constructor = Mnemonica;
-
-	// instance of self Constructor type
-	odp(Mnemonica.prototype, Symbol.hasInstance, {
-		get() {
-			return getTypeChecker(this.constructor.name);
-		}
-	});
-
-	odp(Mnemonica.prototype, SymbolGaia, {
-		get() {
-			return gaia;
-		}
-	});
-
-	odp(Mnemonica.prototype, SymbolConstructorName, {
-		get() {
-			return MNEMONICA;
-		}
-	});
-
-	odp(Mnemonica.prototype, 'extract', {
-		get() {
-			return function () {
-				return extract(this);
-			};
-		}
-	});
-
-	odp(Mnemonica.prototype, 'pick', {
-		get() {
-			return function (...args) {
-				return pick(this, ...args);
-			};
-		}
-	});
-
-	odp(Mnemonica.prototype, 'parent', {
-		get() {
-			return function (constructorLookupPath) {
-				return parent(this, constructorLookupPath);
-			};
-		}
-	});
-
-	odp(Mnemonica.prototype, 'clone', {
-		get() {
-			return this.fork();
-		}
-	});
-
-	odp(Mnemonica.prototype, 'fork', {
-		get() {
-
-			const {
-				__type__: type,
-				__collection__: collection,
-				__parent__: existentInstance,
-				__args__,
-				__self__
-			} = this;
-
-			const {
-				isSubType,
-				TypeName
-			} = type;
-
-			// 'function', cause might be called with 'new'
-			return function (...forkArgs) {
-
-				var forked;
-				const Constructor = isSubType ?
-					existentInstance :
-					collection;
-
-				const args = forkArgs.length ? forkArgs : __args__;
-				
-
-				if (this === __self__) {
-					forked = new (Constructor[TypeName])(...args);
-				} else {
-					// fork.call ? let's do it !
-					forked = createInstance.call(this, type, args);
-				}
-
-				return forked;
-
-			};
-		}
-	});
-
-	const proto = new Mnemonica();
-
-	Object.setPrototypeOf(this, proto);
-
-};
-
-const addThen = (inheritedInstance, then) => {
-	inheritedInstance = inheritedInstance
-		.then((instance) => {
-			const finish = createInstance.call(
-				instance,
-				then.subtype,
-				then.args
-			);
-			return finish;
-		});
-	return inheritedInstance;
-};
-
-const makeWaiter = (inheritedInstance, type, then) => {
-
-	inheritedInstance = inheritedInstance
-		.then((instance) => {
-			if (instance.__self__ !== instance) {
-				// it was async instance,
-				// so we have to add all the stuff
-				// for sync instances it was done already
-				postProcessing(instance, type.subtypes);
-			}
-			return instance;
-		});
-
-	if (then) {
-		inheritedInstance = addThen(inheritedInstance, then);
-	}
-
-	Object.entries(type.subtypes).forEach((entry) => {
-		const [name, subtype] = entry;
-		inheritedInstance[name] = (...args) => {
-			const thens = {
-				name,
-				subtype,
-				args,
-			};
-			inheritedInstance = makeWaiter(inheritedInstance, subtype, thens);
-			return inheritedInstance;
-		};
-	});
-
-	return inheritedInstance;
-
-};
-
-const postProcessing = (inheritedInstance, subtypes) => {
-
-	odp(inheritedInstance, '__self__', {
-		get() {
-			return inheritedInstance;
-		}
-	});
-
-	addSubTypes(inheritedInstance, subtypes);
-	invokePostHooks(inheritedInstance);
-
-};
-
-
-const createInstance = function (type, args) {
-
-	const {
-		constructHandler,
-		proto,
-		useOldStyle,
-		subtypes,
-		TypeName
-	} = type;
-
-	const existentInstance = this;
-
-	const ModificationConstructor = getModificationConstructor(useOldStyle);
-
-	const ModificatorType = constructHandler();
-
-	const opts = {
-		type,
-		existentInstance,
-		get args() {
-			return args;
-		},
-	};
-
-	const InstanceModificator = ModificationConstructor.call(
-		existentInstance,
-		ModificatorType,
-		Object.assign({}, proto),
-		addProps.bind(opts)
-	);
-
-	opts.InstanceModificator = InstanceModificator;
-
-	invokePreHooks(opts);
-	const inheritedInstance = proceedInheritance(opts);
-
-	if (inheritedInstance instanceof Promise) {
-
-		const waiter = makeWaiter(inheritedInstance, type);
-
-		odp(waiter, SymbolConstructorName, {
-			get() {
-				return TypeName;
-			}
-		});
-
-		return waiter;
-
-	}
-
-	postProcessing(inheritedInstance, subtypes);
-
-	return inheritedInstance;
-
-};
-
-const Gaia = function (Uranus) {
-
-	const gaia = !Uranus ? this :
-		Object.setPrototypeOf(this, Uranus);
-
-	odp(gaia, MNEMONICA, {
-		get() {
-			return !Uranus ? GAIA : URANUS;
-		}
-	});
-
-	return gaia;
-};
-
-
-const TypeProxy = function (type, Uranus) {
-
-	const {
-		namespace,
-		isSubType,
-		constructHandler
-	} = type;
-
-	this.__type__ = type;
-
-	if (!isSubType) {
-
-		odp(this, SymbolGaia, {
-			get() {
-				return new Mnemosyne(namespace, new Gaia(Uranus));
-			}
-		});
-
-	}
-
-	const typeProxy = new Proxy(constructHandler, this);
-
-	return typeProxy;
-
-};
-
-const reProxifyInstance = (type, Uranus, ...args) => {
-	const instance = new (new TypeProxy(type, Uranus))(...args);
-	return instance;
-
-};
-
-TypeProxy.prototype.get = function (target, prop) {
-
-	const {
-		__type__: type
-	} = this;
-
-	if (prop === 'call') {
-		return function (Uranus, ...args) {
-			return reProxifyInstance(type, Uranus, ...args);
-			// return new (new TypeProxy(type, Uranus))(...args);
-		};
-	}
 	
-	if (prop === 'apply') {
-		return function (Uranus, args) {
-			return reProxifyInstance(type, Uranus, ...args);
-			// return new (new TypeProxy(type, Uranus))(...args);
-		};
-	}
-
-	const propDeclaration = type[prop];
-	if (propDeclaration) {
-		return propDeclaration;
-	}
-
-	// used for existent props with value
-	// undefined || null || false 
-	if (type.hasOwnProperty(prop)) {
-		return propDeclaration;
-	}
-
-	// SomeType.SomeSubType
-	if (type.subtypes.hasOwnProperty(prop)) {
-		return type.subtypes[prop];
-	}
-
-	return target[prop];
-
-};
-
-TypeProxy.prototype.construct = function (target, args) {
-	const {
-		__type__: type
-	} = this;
-	const gaia = this[SymbolGaia];
-	const instance = createInstance.call(gaia, type, args);
-	return instance;
-};
-
+const {
+	checkProto,
+	getTypeChecker,
+	CreationHandler
+} =	require('./utils');
 
 const TypeDescriptor = function (
 	types,
