+++ test/index.js
@@ -205,19 +205,21 @@ const AsyncType = define('AsyncType', async function (data) {
 	});
 });
 
-const SubOfAsync = AsyncType.define('SubOfAsync', function (data) {
+AsyncType.SubOfAsync = function (data) {
 	Object.assign(this, {
 		data
 	});
-});
+};
 
-const NestedAsyncType = SubOfAsync.define('NestedAsyncType', async function (data) {
+AsyncType.SubOfAsync.NestedAsyncType = async function (data) {
 	return Object.assign(this, {
 		data
 	});
-}, {
+};
+AsyncType.SubOfAsync.NestedAsyncType.prototype = {
 	description: 'nested async instance'
-});
+};
+const NestedAsyncType = AsyncType.SubOfAsync.NestedAsyncType;
 
 const SubOfNestedAsync = NestedAsyncType.define('SubOfNestedAsync', function (data) {
 	Object.assign(this, {
@@ -811,7 +813,7 @@ describe('Main Test', () => {
 					expect(nestedAsyncInstance).instanceof(AsyncType);
 					expect(nestedAsyncInstance).instanceof(NestedAsyncType);
 					expect(nestedAsyncSub).instanceof(AsyncType);
-					expect(nestedAsyncSub).instanceof(SubOfAsync);
+					expect(nestedAsyncSub).instanceof(AsyncType.SubOfAsync);
 					expect(nestedAsyncSub).instanceof(NestedAsyncType);
 					expect(nestedAsyncSub).instanceof(SubOfNestedAsync);
 					expect(SubOfNestedAsyncPostHookData
